% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alpha.R
\name{full_alpha}
\alias{full_alpha}
\title{Full posterior for Alpha}
\usage{
full_alpha(alpha, L, nL, y, XX, J, delta, C, beta, PHI1, PHI2, gamma)
}
\arguments{
\item{alpha}{The current value of the Alpha parameter.}

\item{L}{The number of levels in the model.}

\item{nL}{A vector containing the length of each level.}

\item{y}{The observed data.}

\item{XX}{The explanatory variable matrix.}

\item{J}{Indicator matrix for zero values in the data.}

\item{delta}{Indicator matrix for censored data.}

\item{C}{The threshold value.}

\item{beta}{The Beta parameter in the model.}

\item{PHI1}{The spatial effects parameter for PHI1.}

\item{PHI2}{The spatial effects parameter for PHI2.}

\item{gamma}{The Gamma parameter in the model.}
}
\value{
The computed posterior probability for Alpha.
}
\description{
This function computes the full posterior probability for the Alpha parameter in a Bayesian model.
}
