% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_beta.R
\name{full_beta}
\alias{full_beta}
\title{Full posterior for Beta}
\usage{
full_beta(beta, L, nL, y, XX, J, delta, C, a, b, alpha, PHI2)
}
\arguments{
\item{beta}{The current value of the Beta parameter.}

\item{L}{The number of levels in the model.}

\item{nL}{A vector containing the length of each level.}

\item{y}{The observed data.}

\item{XX}{The explanatory variable matrix.}

\item{J}{Indicator matrix for zero values in the data.}

\item{delta}{Indicator matrix for censored data.}

\item{C}{The threshold value.}

\item{a}{The shape parameter for the inverse gamma prior.}

\item{b}{The rate parameter for the inverse gamma prior.}

\item{alpha}{The alpha parameter in the model.}

\item{PHI2}{The spatial effects parameter for PHI2.}
}
\value{
The computed posterior probability for Beta.
}
\description{
This function computes the full posterior probability for the Beta parameter in a Bayesian model.
}
