% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gamma.R
\name{full_gamma}
\alias{full_gamma}
\title{Full posterior for Gamma}
\usage{
full_gamma(gamma, L, nL, y, XX, J, delta, C, beta, PHI1, PHI2, alpha)
}
\arguments{
\item{gamma}{The current value of the Gamma parameter.}

\item{L}{The number of levels in the model.}

\item{nL}{A vector containing the length of each level.}

\item{y}{The observed data.}

\item{XX}{The explanatory variable matrix.}

\item{J}{Indicator matrix for zero values in the data.}

\item{delta}{Indicator matrix for censored data.}

\item{C}{The threshold value.}

\item{beta}{The Beta parameter in the model.}

\item{PHI1}{The spatial effects parameter for PHI1.}

\item{PHI2}{The spatial effects parameter for PHI2.}

\item{alpha}{The Alpha parameter in the model.}
}
\value{
The computed posterior probability for Gamma.
}
\description{
This function computes the full posterior probability for the Gamma parameter in a Bayesian model.
}
