% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_PHI.R
\name{full_PHI2}
\alias{full_PHI2}
\title{Full posterior for PHI2}
\usage{
full_PHI2(
  PHI2,
  L,
  nL,
  y,
  XX,
  J,
  delta,
  C,
  beta,
  PHI1,
  alpha,
  gamma,
  rho,
  SIGMA,
  MA
)
}
\arguments{
\item{PHI2}{The current value of the PHI2 parameter.}

\item{L}{The number of levels.}

\item{nL}{The number of samples at each level.}

\item{y}{The observed data.}

\item{XX}{The explanatory variable matrix.}

\item{J}{Indicator matrix for zero values.}

\item{delta}{Indicator matrix for censored data.}

\item{C}{Threshold value.}

\item{beta}{The Beta parameter in the model.}

\item{PHI1}{The PHI1 parameter.}

\item{alpha}{The Alpha parameter.}

\item{gamma}{The Gamma parameter.}

\item{rho}{Spatial smoothing parameter.}

\item{SIGMA}{The covariance matrix.}

\item{MA}{The matrix of neighbors.}
}
\value{
The posterior probability for PHI2.
}
\description{
Computes the full posterior probability for PHI2 in the model.
}
