% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_SIGMA.R
\name{full_SIGMA}
\alias{full_SIGMA}
\title{Full posterior for SIGMA}
\usage{
full_SIGMA(PHI, L, rho, nu0, S0, sigma)
}
\arguments{
\item{PHI}{The spatial effects parameter for PHI.}

\item{L}{The number of levels in the model.}

\item{rho}{Spatial smoothing parameter.}

\item{nu0}{Degrees of freedom for the Wishart prior.}

\item{S0}{Scale matrix for the Wishart prior.}

\item{sigma}{The covariance matrix.}
}
\value{
The computed posterior for SIGMA.
}
\description{
Computes the full posterior for the SIGMA parameter in the model.
}
