% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_hastings.R
\name{sim.theta}
\alias{sim.theta}
\title{Metropolis-Hastings Algorithm for Bayesian Inference}
\usage{
sim.theta(nsim, burn, thin, L, nL)
}
\arguments{
\item{nsim}{The number of simulations to run.}

\item{burn}{The number of burn-in iterations.}

\item{thin}{The thinning interval.}

\item{L}{The number of levels in the model.}

\item{nL}{A vector of the number of samples at each level.}
}
\value{
A matrix containing the sampled values for the parameters.
}
\description{
This function performs the Metropolis-Hastings algorithm for Bayesian inference on the model.
}
