% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_J_delta.R
\name{y_J_delta}
\alias{y_J_delta}
\title{Generate Data}
\usage{
y_J_delta(L, nL, beta)
}
\arguments{
\item{L}{Number of levels.}

\item{nL}{A vector specifying the number of samples at each level.}

\item{beta}{A vector of beta values.}
}
\value{
A list containing various generated data objects including y0, y, delta, and others.
}
\description{
This function generates data for a Bayesian model with spatial effects and a zero-inflated discrete weibull distribution.
}
